topology_builder/requirements.txt
Babel==2.12.1
basemap==1.3.8
basemap-data==1.3.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
cycler==0.11.0
distro==1.8.0
fonttools==4.42.1
ghp-import==2.1.0
griffe==0.36.2
idna==3.4
iniconfig==2.0.0
Jinja2==3.1.2
jplephem==2.18
kiwisolver==1.4.5
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.3
mercurial==6.5
mergedeep==1.3.4
mkdocs==1.5.2
mkdocs-autorefs==0.5.0
mkdocs-material==9.3.1
mkdocs-material-extensions==1.1.1
mkdocstrings==0.23.0
mkdocstrings-python==1.6.3
networkx==3.1
numpy==1.25.2
packaging==23.1
paginate==0.5.6
pathspec==0.11.2
Pillow==10.0.1
platformdirs==3.10.0
pluggy==1.3.0
Pygments==2.16.1
pymdown-extensions==10.3
pyparsing==3.1.1
pyproj==3.6.1
pyshp==2.3.1
pytest==7.4.2
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
pyyaml_env_tag==0.1
regex==2023.8.8
requests==2.31.0
sgp4==2.22
six==1.16.0
skyfield==1.46
typer==0.9.0
typing_extensions==4.7.1
urllib3==2.0.4
watchdog==3.0.0
